+++ packages/header/src/components/NavigationBurgerMenu.tsx
@@ -55,7 +55,10 @@ const NavigationBurgerMenu = React.forwardRef<HTMLDivElement, NavigationProps>(
         <ul aria-label="Main navigation" role="menubar" tabIndex={0}>
           {extendedNavArray.map((navItem, i) => {
             const { label, subNavGroups } = navItem;
-            const id = `nav-item-${label.replace(/^[^a-z]+|[^\w:.-]+/gi, '')}`;
+            const id = `burger-nav-item-${label.replace(
+              /^[^a-z]+|[^\w:.-]+/gi,
+              '',
+            )}`;
             const hasSubNav = subNavGroups && subNavGroups?.length > 0;
             const isActive = activeNavItem === id;
             const isLastItem = i + 1 === navigation.length;
@@ -71,6 +74,7 @@ const NavigationBurgerMenu = React.forwardRef<HTMLDivElement, NavigationProps>(
                   navItem.isBurgerMenuOnly ? styles.headerItems : ''
                 }`}
                 key={`burger-${label}`}
+                role="none"
               >
                 {hasSubNav ? (
                   <>
@@ -80,8 +84,16 @@ const NavigationBurgerMenu = React.forwardRef<HTMLDivElement, NavigationProps>(
                       aria-haspopup={true}
                       aria-label={`${label} - ${locale.spaceBarNotification}`}
                       onClick={() =>
-                        itemOnClick(`${label}`, trackingFn, () =>
-                          toggleSubNav(id, isActive),
+                        itemOnClick(
+                          {
+                            fn: trackingFn,
+                            obj: {
+                              label: `${label}`,
+                              type: 'nav_item',
+                              navType: 'burger_menu',
+                            },
+                          },
+                          () => toggleSubNav(id, isActive),
                         )
                       }
                       onKeyDown={e => keyboardOpen(e, id, isActive)}
@@ -94,6 +106,7 @@ const NavigationBurgerMenu = React.forwardRef<HTMLDivElement, NavigationProps>(
                       aria-label={`${label} ${locale.subMenuLabel}`}
                       className={styles.collapse}
                       open={isActive}
+                      role="group"
                     >
                       <SubNav
                         Link={Link}
@@ -112,7 +125,18 @@ const NavigationBurgerMenu = React.forwardRef<HTMLDivElement, NavigationProps>(
                     {...commonProps}
                     href={navItem.href}
                     onClick={() =>
-                      itemOnClick(`${label}`, trackingFn, navItem.onClick)
+                      itemOnClick(
+                        {
+                          fn: trackingFn,
+                          obj: {
+                            label: `${label}`,
+                            type: 'nav_item',
+                            href: navItem.href,
+                            navType: 'burger_menu',
+                          },
+                        },
+                        navItem.onClick,
+                      )
                     }
                   >
                     <Typography variant="subheading3">{label}</Typography>
@@ -132,7 +156,22 @@ const NavigationBurgerMenu = React.forwardRef<HTMLDivElement, NavigationProps>(
                 <Typography Component="span" variant="caption2">
                   {aux.tel.label || locale.auxTelLabel}
                 </Typography>
-                <a href={`tel:${aux.tel.value}`}>{aux.tel.value}</a>
+                <a
+                  href={`tel:${aux.tel.value}`}
+                  onClick={() =>
+                    itemOnClick({
+                      fn: trackingFn,
+                      obj: {
+                        label: 'aux telephone',
+                        type: 'nav_item',
+                        href: `tel:${aux.tel?.value}`,
+                        navType: 'burger_menu',
+                      },
+                    })
+                  }
+                >
+                  {aux.tel.value}
+                </a>
               </p>
             )}
             {aux?.email && (
@@ -140,7 +179,22 @@ const NavigationBurgerMenu = React.forwardRef<HTMLDivElement, NavigationProps>(
                 <Typography Component="span" variant="caption2">
                   {aux.email.label || locale.auxEmailLabel}
                 </Typography>
-                <a href={`mailto:${aux.email.value}`}>{aux.email.value}</a>
+                <a
+                  href={`mailto:${aux.email.value}`}
+                  onClick={() =>
+                    itemOnClick({
+                      fn: trackingFn,
+                      obj: {
+                        label: 'aux email',
+                        type: 'nav_item',
+                        href: `tel:${aux.email?.value}`,
+                        navType: 'burger_menu',
+                      },
+                    })
+                  }
+                >
+                  {aux.email.value}
+                </a>
               </p>
             )}
           </aside>